bmp: use double variant of ceil func
We use ceill and ceil in this code, but it's not clear we need the long
double variant of ceill here.  The input multiply is already done with
double precision (the 0.1 literal is a double), and not all C libraries
offer long double variants.  Change to ceil and see if anyone notices.

Closes issue #123.

+++ src/gd_bmp.c
@@ -792,7 +792,7 @@ static int bmp_read_1bit(gdImagePtr im, gdIOCtxPtr infile, bmp_info_t *info, bmp
 	}
 
 	/* The line must be divisible by 4, else its padded with NULLs */
-	padding = ((int)ceill(0.1 * info->width)) % 4;
+	padding = ((int)ceil(0.1 * info->width)) % 4;
 	if (padding) {
 		padding = 4 - padding;
 	}
